@@ -8,7 +8,8 @@
   "scripts": {
     "copy-package": "shx cp main/typescript/package.json build/dist/package.json",
     "clean": "del-cli -f build/dist ./*.tgz",
-    "build": "yarn upgrade --pattern strata && tsc && cd main/typescript && yarn run increment-snapshot && cd ../.. && yarn run copy-package && cd build/dist && yarn pack",
+    "build-init": "yarn run clean && yarn install && yarn upgrade --pattern strata",
+    "build": "yarn run build-init && tsc && cd main/typescript && yarn run increment-snapshot && cd ../.. && yarn run copy-package && cd build/dist && yarn pack",
     "publish-package": "yarn run build && cd build/dist && yarn publish --non-interactive",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
@@ -18,10 +19,10 @@
   "dependencies": {
     "guid-typescript": "^1.0.9",
     "rxjs": "^6.5.5",
-    "strata.foundation.core": "^0.0.0-snapshot.18",
-    "typescript": "^5.1.6"
+    "strata.foundation.core": "^0.0.0-snapshot.28"
   },
   "devDependencies": {
-    "shx": "^0.3.4"
+    "shx": "^0.3.4",
+    "typescript": "^5.2.0"
   }
 }
